Adding Hough
To create a video with the lines drawn. Notice that there is a
particular function "zbufer_cdata" that is needed to get the figure
hwere we have drawn the liens without popping out any image.
In this file you can find the function to manage the video :) there are
a lot of more way to do the same..